# api/openapi.yaml
openapi: 3.0.0
info:
  title: SHRTIC API
  description: |
    Сервис сокращения ссылок с аналитикой переходов.
    
    Аутентификация:
    - Пользователь регистрируется и входит, получает JWT токен
    - Токен передаётся через cookie (shrtic_token) или в заголовке Authorization: Bearer <token>
    - Токен живёт 24 часа
    
    Сокращение ссылок:
    - Создать короткую ссылку может любой, даже без авторизации
    - Если пользователь авторизован, ссылка привязывается к нему
    - Просмотр истории и статистики — только для авторизованных
    - При переходе по короткой ссылке автоматически собирается статистика
    
    Статистика:
    - Браузеры (Chrome, Firefox, Safari, Edge, Other)
    - Устройства (Desktop, Mobile, Tablet)
    - Страны (определяется по IP через ip-api.com)
    - Источники перехода (заголовок Referer)

    Типы ошибок:
    - validation_error — ошибка валидации входных данных (400)
    - invalid_credentials — неверный логин или пароль (401)
    - unauthorized — не авторизован (401)
    - forbidden — нет прав доступа (403)
    - not_found — ресурс не найден (404)
    - conflict — конфликт данных (409)
    - rate_limit — слишком много попыток входа (429)
    - internal_error — внутренняя ошибка сервера (500)
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

tags:
  - name: Auth
    description: Аутентификация и управление пользователями
  - name: Links
    description: Сокращение, управление и статистика ссылок
  - name: Redirect
    description: Переход по коротким ссылкам
  - name: Web
    description: HTML страницы
  - name: System
    description: Системные эндпоинты

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен, полученный при входе.
        Пример: Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        Также принимается через cookie shrtic_token

  schemas:

    RegisterRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 50
          description: Уникальный логин пользователя
          example: "username"
        password:
          type: string
          minLength: 6
          maxLength: 100
          format: password
          description: Пароль, будет захэширован через bcrypt
          example: "secret123"

    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          description: Логин пользователя
          example: "username"
        password:
          type: string
          format: password
          description: Пароль
          example: "secret123"

    LoginResponse:
      type: object
      properties:
        role:
          type: string
          enum: [user]
          description: Роль пользователя
          example: "user"
      description: "Токен передаётся в cookie shrtic_token, не в JSON"

    MeResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: "username"
        created_at:
          type: string
          format: date-time
          example: "2026-04-26T20:30:00Z"

    CreateLinkRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Длинный URL для сокращения (https:// добавится автоматически)
          example: "github.com/username/repo"

    Link:
      type: object
      description: Информация о ссылке
      properties:
        id:
          type: integer
          description: Уникальный идентификатор
          example: 42
        short_url:
          type: string
          description: Относительный путь короткой ссылки
          example: "/r/x7kQ2"
        long_url:
          type: string
          description: Оригинальный длинный URL
          example: "https://github.com/username/repo"
        clicks:
          type: integer
          description: Общее количество переходов
          example: 1247
        created_at:
          type: string
          format: date-time
          description: Дата и время создания
          example: "2026-04-26T20:30:00Z"

    StatItem:
      type: object
      properties:
        name:
          type: string
          description: Название (Chrome, Desktop, Россия, Twitter)
          example: "Chrome"
        percentage:
          type: number
          format: double
          description: Процент от общего числа переходов
          example: 62.5
        count:
          type: integer
          description: Абсолютное количество переходов
          example: 780

    LinkStats:
      type: object
      description: Полная статистика по ссылке
      properties:
        total_clicks:
          type: integer
          description: Общее количество переходов
          example: 1247
        browsers:
          type: array
          description: Распределение по браузерам
          items:
            $ref: '#/components/schemas/StatItem'
        devices:
          type: array
          description: Распределение по типам устройств
          items:
            $ref: '#/components/schemas/StatItem'
        countries:
          type: array
          description: Распределение по странам
          items:
            $ref: '#/components/schemas/StatItem'
        referrers:
          type: array
          description: Распределение по источникам перехода
          items:
            $ref: '#/components/schemas/StatItem'

    IdResponse:
      type: object
      properties:
        id:
          type: integer
          description: "Всегда 0 (ID не возвращается, только статус 201)"
          example: 0

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Информационное сообщение
          example: "Операция выполнена успешно"

    ErrorResponse:
      type: object
      description: Стандартный ответ с ошибкой
      properties:
        error:
          type: string
          description: Тип ошибки
          example: "validation_error"
        message:
          type: string
          description: Человекочитаемое сообщение
          example: "Поле url обязательно"
        details:
          type: object
          additionalProperties: true
          description: Дополнительные детали ошибки
          example:
            blocked_for_seconds: 900

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"

paths:

  /:
    get:
      summary: Главная страница
      description: HTML страница с формой сокращения ссылок
      tags: [Web]
      responses:
        '200':
          description: HTML страница
          content:
            text/html: {}

  /login:
    get:
      summary: Страница входа и регистрации
      tags: [Web]
      responses:
        '200':
          description: HTML страница
          content:
            text/html: {}

  /cabinet:
    get:
      summary: Личный кабинет
      description: Требует авторизации. Без токена редирект на /login
      tags: [Web]
      responses:
        '200':
          description: HTML страница
          content:
            text/html: {}
        '302':
          description: Редирект на /login

  /static/{file}:
    get:
      summary: Статические файлы
      description: CSS, JS, изображения
      tags: [Web]
      parameters:
        - name: file
          in: path
          required: true
          schema:
            type: string
          example: "app.css"
      responses:
        '200':
          description: Файл
          content:
            text/css: {}
            application/javascript: {}


  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Сервер работает
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "ok"

  /metrics:
    get:
      summary: Prometheus метрики
      description: Метрики для мониторинга (RPS, latency, активные ссылки, пользователи)
      tags: [System]
      responses:
        '200':
          description: Метрики в формате Prometheus
          content:
            text/plain: {}

  /api/auth/register:
    post:
      summary: Регистрация нового пользователя
      description: |
        Создаёт новую учётную запись.
        Доступно без авторизации.
      operationId: register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
              example:
                id: 0
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Вход в систему
      description: |
        Аутентификация пользователя.
        При успехе JWT токен устанавливается в cookie shrtic_token.
        Rate limiting: 5 попыток за 15 минут, затем блокировка на 15 минут.
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                role: "user"
          headers:
            Set-Cookie:
              schema:
                type: string
              description: "shrtic_token=...; HttpOnly; Path=/; Max-Age=86400"
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Слишком много попыток входа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      summary: Выход из системы
      description: Удаляет cookie с токеном
      operationId: logout
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Выход выполнен успешно"

  /api/auth/me:
    get:
      summary: Информация о текущем пользователе
      description: Возвращает данные авторизованного пользователя
      operationId: getMe
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
              example:
                id: 1
                login: "username"
                created_at: "2026-04-26T20:30:00Z"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/links:
    post:
      summary: Сократить ссылку
      description: |
        Создаёт короткую ссылку.
        Доступно без авторизации. 
        Если передан JWT токен, ссылка привязывается к пользователю.
      operationId: createLink
      tags: [Links]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkRequest'
      responses:
        '201':
          description: Ссылка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
              example:
                id: 42
                short_url: "/r/x7kQ2"
                long_url: "https://github.com/username/repo"
                clicks: 0
                created_at: "2026-04-26T20:30:00Z"
        '400':
          description: Невалидный URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Список ссылок пользователя
      description: |
        Возвращает все ссылки, созданные авторизованным пользователем.
        Поддерживает поиск по оригинальному URL.
        Требует авторизации.
      operationId: listLinks
      tags: [Links]
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Поиск по оригинальному URL (частичное совпадение, регистронезависимый)
          example: "github"
      responses:
        '200':
          description: Список ссылок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/links/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ссылки
        schema:
          type: integer
          minimum: 1
    delete:
      summary: Удалить ссылку
      description: Удаляет ссылку и всю связанную с ней статистику
      operationId: deleteLink
      tags: [Links]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ссылка удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Ссылка удалена"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нельзя удалить чужую ссылку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ссылка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/links/{id}/stats:
    parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ссылки
        schema:
          type: integer
          minimum: 1
    get:
      summary: Статистика по ссылке
      description: |
        Возвращает агрегированную статистику переходов:
        браузеры, устройства, страны, источники перехода.
        Требует авторизации и прав на ссылку.
      operationId: getLinkStats
      tags: [Links]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Статистика по ссылке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkStats'
              example:
                total_clicks: 1247
                browsers:
                  - name: "Chrome"
                    percentage: 62.5
                    count: 780
                devices:
                  - name: "Desktop"
                    percentage: 71.0
                    count: 885
                countries:
                  - name: "Россия"
                    percentage: 45.0
                    count: 561
                referrers:
                  - name: "Прямой"
                    percentage: 45.0
                    count: 561
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Нет прав на просмотр статистики чужой ссылки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ссылка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /r/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: Короткий код ссылки (7 символов a-zA-Z0-9)
        schema:
          type: string
          pattern: "^[a-zA-Z0-9]{7}$"
          example: "x7kQ2"
    get:
      summary: Переход по короткой ссылке
      description: |
        Перенаправляет (302 Found) на оригинальный URL.
        При переходе автоматически собирается статистика:
        браузер, устройство, страна, источник.
      operationId: redirect
      tags: [Redirect]
      responses:
        '302':
          description: Перенаправление на оригинальный URL
          headers:
            Location:
              description: Оригинальный URL
              schema:
                type: string
                example: "https://github.com/username/repo"
        '404':
          description: Ссылка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'